Upgrade Dependencies and Example App for Full Android 14 and AGP 8 Support
Upgraded all relevant dependencies and updated the example application to fully support Android 14. With the release of Android 14, it became necessary to update our project to ensure compatibility and to take advantage of the new features and security enhancements that come with the latest Android version. This update helps to maintain optimal performance and user experience on newer devices.