Configure JDK 11 for JitPack Compilation with jitpack.yml
Added a jitpack.yml file to the project to specify the use of Java JDK 11 for compiling the project on JitPack. This change is necessary, since the last Android SDK 34 and AGP 8 are requiring features that are specific to Java JDK 11, which is not the default JDK version used by JitPack. By explicitly setting the JDK version in the jitpack.yml file, we prevent compilation issues and ensure that builds are generated correctly on the JitPack platform.